refactor: smallmols
