+++ src/plinder/eval/docking/stratify_test_set.py
@@ -7,12 +7,17 @@ from pathlib import Path
 
 import numpy as np
 import pandas as pd
+from cloudpathlib import AnyPath
 from tqdm import tqdm
 
+import plinder.core.utils.config
 from plinder.core.scores.index import query_index
 from plinder.core.scores.protein import cross_similarity as protein_cross_similarity
+from plinder.core.structure import smallmols
 from plinder.core.utils.log import setup_logger
-from plinder.data import smallmolecules
+
+cfg = plinder.core.get_config()
+mmp_path = AnyPath(f"{cfg.data.plinder_remote}/mmp/plinder_mms.csv.gz")
 
 LOG = setup_logger(__name__)
 
@@ -67,20 +72,27 @@ def compute_ligand_max_similarities(
 ) -> None:
     if "fp" not in df.columns:
         smiles_fp_dict = {
-            smi: smallmolecules.mol2morgan_fp(smi)
+            smi: smallmols.mol2morgan_fp(smi)
             for smi in df["ligand_rdkit_canonical_smiles"].drop_duplicates().to_list()
         }
         df["fp"] = df["ligand_rdkit_canonical_smiles"].map(smiles_fp_dict)
 
     df_test = df.loc[df[split_label] == test_label][["system_id", "fp"]].copy()
 
-    df_test["tanimoto_similarity_max"] = smallmolecules.tanimoto_maxsim_matrix(
+    df_test["tanimoto_similarity_max"] = smallmols.tanimoto_maxsim_matrix(
         df.loc[df[split_label] == train_label]["fp"].to_list(),
         df_test["fp"].to_list(),
     )
     df_test.drop("fp", axis=1).groupby("system_id").agg("max").reset_index().to_parquet(
         output_file, index=False
     )
+
+
+def compute_ligand_max_mmp_similarities(
+    df: pd.DataFrame, train_label: str, test_label: str, output_file: Path
+) -> None:
+    # TODO: add this!
+    return
 
 
 @dataclass
